Time: 0 ms (100%), Space: 10.2 MB (72.97%) - LeetHub

+++ 1079-sum-of-root-to-leaf-binary-numbers/1079-sum-of-root-to-leaf-binary-numbers.c
@@ -6,7 +6,7 @@
  *     struct TreeNode *right;
  * };
  */
-int preorder(struct TreeNode* root,long long sum){
+int preorder(struct TreeNode* root,int sum){
     if(root==NULL) return 0;
     sum=sum*2+root->val;
     if(root->left==NULL&&root->right==NULL){
